// src/models/home.model.ts
export interface CounterModel {
    title_en: string;
    title_ar: string;
    value: string;
    _id: string;
}



export interface TestimonialModel {
    description_ar: string;
    description_en: string;
    image: string;
    _id: string;
}


export interface PartnerModel {
    image: string;
    _id: string;
}


export interface HomeAboutModel {
    title_en: string;
    title_ar: string;
    description_ar: string;
    description_en: string;
    type: number;
    _id: string;
}